Initial support for new options page added.
The new options page required as part of the change for #3 has been
partially implemented and now allows the dynamic creation and storage of
the settings. Some basic support for multis has been added but this is
not yet working. Singles now works in both page load and mutation modes.