Updated to handle text blocks being updated
This is a fix for #23. It appears the user was experiencing this issue
because when the mutation occurred on the page, the affected node was
the node in question. To get around this I have had to ensure that the
current node is added to the list to be filtered, and then use the
parent if it appears that node is a text node. After speaking to the
affected user this appears to have fixed the issue.